requirements.txt

# --- Realtime analysis ---
numpy==1.26.4
librosa==0.10.2.post1
sounddevice==0.4.6
soundfile==0.12.1
chord-extractor==0.1.3
vamp==1.1.0
requests==2.31.0
python-dotenv==1.0.1

# --- MIDI tooling ---
pretty_midi==0.2.10
mido==1.3.2
miditoolkit==0.1.16
miditok>=3.0.0

# --- Training stack (token LM) ---
torch>=2.3,<3.0
transformers>=4.41,<5.0
datasets>=2.18.0,<3.0
accelerate>=0.31.0,<1.0
peft>=0.11.0,<1.0
trl>=0.9.6,<1.0
tqdm>=4.66.0,<5.0
safetensors>=0.4.2,<1.0
